moving CustomConfig to rpc-client
